style: created a styles.ts for the Login page
