+++ App.tsx
@@ -1,6 +1,6 @@
-import React, {useEffect, useState} from 'react';
+import React, { useEffect, useState } from 'react';
 import { NavigationContainer } from '@react-navigation/native';
-import Login from './app/screens/Login';
+import Login from './app/screens/Onboarding/Login';
 import AppContent from './app/screens/AppContent';
 import { createNativeStackNavigator } from '@react-navigation/native-stack';
 import { User, onAuthStateChanged } from 'firebase/auth';
@@ -10,7 +10,7 @@ import { FIREBASE_AUTH } from './firebaseConfig';
 const Stack = createNativeStackNavigator();
 const AppContentStack = createNativeStackNavigator();
 
-function AppLayout () {
+function AppLayout() {
   return (
     <AppContentStack.Navigator>
       <AppContentStack.Screen name="AppContent" component={AppContent} />
@@ -27,15 +27,15 @@ export default function App() {
       setUser(user);
     });
   }
-  , []);
+    , []);
 
   return (
     <NavigationContainer >
       <Stack.Navigator initialRouteName='Login'>
         {user ? (
-          <Stack.Screen name="AppLayout" component={AppLayout} options={{headerShown: false}} />
+          <Stack.Screen name="AppLayout" component={AppLayout} options={{ headerShown: false }} />
         ) : (
-          <Stack.Screen name="Login" component={Login} options={{headerShown: false}}/>
+          <Stack.Screen name="Login" component={Login} options={{ headerShown: false }} />
         )}
       </Stack.Navigator>
     </NavigationContainer>
