+++ app/screens/Onboarding/styles.ts
@@ -1,0 +1,18 @@
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+    container: {
+        alignItems: 'center',
+        flex: 1,
+        justifyContent: 'center',
+    },
+    input: {
+        borderRadius: 5,
+        borderWidth: 1,
+        padding: 10,
+        width: '80%',
+    },
+
+});
+
+export default styles;
